+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - helper [zip](src/helpers/files/zip.ts)
 - helper [exportToExcel](src/helpers/files/exportToExcel.ts)
 - helper [saveAsFile](src/helpers/files/saveAsFile.ts)
+- helper [csv](src/helpers/files/csv.ts)
 - [SortElement](src/sortElement.ts)
 
 ## 1.2.7 (Aug 25, 2026)